requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyGithub
python-gitlab
google-generativeai
jinja2